% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_get_holidays}
\alias{api_get_holidays}
\title{Get holidays from the different years}
\usage{
api_get_holidays(start_date, end_date, request_country = "FR")
}
\arguments{
\item{start_date}{\code{(string)} starting date, format : yyyy/mm/dd}

\item{end_date}{\code{(string)} end_date, format : yyyy/mm/dd}

\item{request_country}{\code{(string)}  country code of the country requested. Example : 'FR'}
}
\value{
holidays \code{(vector)}
}
\description{
Get holidays from the different years
}
