% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_get_weather_enedis}
\alias{api_get_weather_enedis}
\title{Process weather data from ENEDIS API}
\usage{
api_get_weather_enedis(start_date, end_date)
}
\arguments{
\item{start_date}{\code{(string)} starting date, format : yyyy-mm-dd}

\item{end_date}{\code{(string)} end_date, format : yyyy-mm-dd}
}
\description{
Process weather data from ENEDIS API
}
