% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_process_weather_enedis}
\alias{api_process_weather_enedis}
\title{Process weather data from ENEDIS API}
\usage{
api_process_weather_enedis(df_weather, start_date, end_date)
}
\arguments{
\item{df_weather}{\code{(data.table)} Data obtained after api_request_weather_enedis}

\item{start_date}{\code{(string)} starting date, format : yyyy/mm/dd}

\item{end_date}{\code{(string)} end_date, format : yyyy/mm/dd}
}
\value{
None
}
\description{
Process weather data from ENEDIS API
}
