% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_request_holidays}
\alias{api_request_holidays}
\title{Request Nager API for holidays data}
\usage{
api_request_holidays(request_year, request_country)
}
\value{
holidays (dataframe) for a year and a country
}
\description{
Request Nager API for holidays data
}
