% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_request_weather_enedis}
\alias{api_request_weather_enedis}
\title{Request ENEDIS API for weather data}
\usage{
api_request_weather_enedis()
}
\value{
df_weather \code{(data.table)} temperature data from enedis api
}
\description{
Request ENEDIS API for weather data
}
