% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{camel_to_snake}
\alias{camel_to_snake}
\title{Converts a character from camelCase to snake_case.}
\usage{
camel_to_snake(inputString)
}
\arguments{
\item{inputString}{A character character in camelCase format that needs to be converted to snake_case.}
}
\value{
\code{(string)} String in snake_case format.
}
\description{
This function takes a character formatted in camelCase and converts it to snake_case.
}
