% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatic_corrector.R
\docType{class}
\name{ClimaticCorrector}
\alias{ClimaticCorrector}
\title{ClimaticCorrector}
\description{
An R6 class which stands as the model structure for each climatic correction model
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{timeseries_model}}{\code{(TimeseriesModel)} A TimeseriesModel object instantiated.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClimaticCorrector-new}{\code{ClimaticCorrector$new()}}
\item \href{#method-ClimaticCorrector-fit}{\code{ClimaticCorrector$fit()}}
\item \href{#method-ClimaticCorrector-apply}{\code{ClimaticCorrector$apply()}}
\item \href{#method-ClimaticCorrector-get_operator}{\code{ClimaticCorrector$get_operator()}}
\item \href{#method-ClimaticCorrector-export}{\code{ClimaticCorrector$export()}}
\item \href{#method-ClimaticCorrector-clone}{\code{ClimaticCorrector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClimaticCorrector-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClimaticCorrector-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClimaticCorrector$new(timeseries_model, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClimaticCorrector-fit"></a>}}
\if{latex}{\out{\hypertarget{method-ClimaticCorrector-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the model using the TimeseriesModel object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClimaticCorrector$fit(timeseries, weather_observed, fold_varname = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} The output/response data}

\item{\code{weather_observed}}{\code{(data.table | TimeseriesDT)} The input weather observed}

\item{\code{fold_varname}}{\code{(string)} Default NULL. The variable name to use to make a cross validation. Only used for a StdModel and Operator2Moments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClimaticCorrector-apply"></a>}}
\if{latex}{\out{\hypertarget{method-ClimaticCorrector-apply}{}}}
\subsection{Method \code{apply()}}{
Apply the model and make the climatic correction using an Operator object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClimaticCorrector$apply(timeseries, weather_observed, weather_target)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} The output/response data}

\item{\code{weather_observed}}{\code{(data.table | TimeseriesDT)} The input weather observed}

\item{\code{weather_target}}{\code{(data.table | TimeseriesDT)} The input weather target}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(TimeseriesDT)} The output timeseries corrected from the weather target
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClimaticCorrector-get_operator"></a>}}
\if{latex}{\out{\hypertarget{method-ClimaticCorrector-get_operator}{}}}
\subsection{Method \code{get_operator()}}{
Get the operator of the Climatic Corrector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClimaticCorrector$get_operator()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(Operator) The Operator of the Climatic Corrector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClimaticCorrector-export"></a>}}
\if{latex}{\out{\hypertarget{method-ClimaticCorrector-export}{}}}
\subsection{Method \code{export()}}{
Export the ClimaticCorrector in RDS.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClimaticCorrector$export(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{\code{(string)} Path file to write the rds file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClimaticCorrector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClimaticCorrector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClimaticCorrector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
