% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat.R
\docType{class}
\name{CopyCat}
\alias{CopyCat}
\title{Copycat}
\description{
Performs a Knn - like model.
}
\section{Super class}{
\code{\link[CorrClim:TimeseriesModel]{CorrClim::TimeseriesModel}} -> \code{CopyCat}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CopyCat-new}{\code{CopyCat$new()}}
\item \href{#method-CopyCat-print}{\code{CopyCat$print()}}
\item \href{#method-CopyCat-fit_fun}{\code{CopyCat$fit_fun()}}
\item \href{#method-CopyCat-predict_fun}{\code{CopyCat$predict_fun()}}
\item \href{#method-CopyCat-clone}{\code{CopyCat$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="check_timeseries"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-check_timeseries'><code>CorrClim::TimeseriesModel$check_timeseries()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="cv_predict"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-cv_predict'><code>CorrClim::TimeseriesModel$cv_predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="export"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-export'><code>CorrClim::TimeseriesModel$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="fit"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-fit'><code>CorrClim::TimeseriesModel$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_formula"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_formula'><code>CorrClim::TimeseriesModel$get_formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_model"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_model'><code>CorrClim::TimeseriesModel$get_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_smoothers"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_smoothers'><code>CorrClim::TimeseriesModel$get_smoothers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="predict"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-predict'><code>CorrClim::TimeseriesModel$predict()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CopyCat-new"></a>}}
\if{latex}{\out{\hypertarget{method-CopyCat-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CopyCat$new(formula = "y ~ temperature", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CopyCat-print"></a>}}
\if{latex}{\out{\hypertarget{method-CopyCat-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CopyCat$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CopyCat-fit_fun"></a>}}
\if{latex}{\out{\hypertarget{method-CopyCat-fit_fun}{}}}
\subsection{Method \code{fit_fun()}}{
Fit function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CopyCat$fit_fun(model, X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{\code{(Any)}}

\item{\code{X}}{\code{(data.table | TimeseriesDT)} Timeseries data to fit. Should contains all variables in the formula}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CopyCat-predict_fun"></a>}}
\if{latex}{\out{\hypertarget{method-CopyCat-predict_fun}{}}}
\subsection{Method \code{predict_fun()}}{
Predict function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CopyCat$predict_fun(model, X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{\code{(Any)}}

\item{\code{X}}{\code{(data.table | TimeseriesDT)} Timeseries data to predict.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CopyCat-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CopyCat-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CopyCat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
