% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correlation_score}
\alias{correlation_score}
\title{Compute correlation of a 2 columns data.table}
\usage{
correlation_score(ts)
}
\arguments{
\item{ts}{\code{(data.table)} The data table to compute the correlation on}
}
\value{
\code{(float)} The correlation between the two columns
}
\description{
Compute correlation of a 2 columns data.table
}
