% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\docType{class}
\name{DummySmoother}
\alias{DummySmoother}
\title{Dummy Smoother}
\description{
Dummy smoother. Performs nothing.
}
\section{Super class}{
\code{\link[CorrClim:Smoother]{CorrClim::Smoother}} -> \code{DummySmoother}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DummySmoother-new}{\code{DummySmoother$new()}}
\item \href{#method-DummySmoother-fit_fun}{\code{DummySmoother$fit_fun()}}
\item \href{#method-DummySmoother-smooth_fun}{\code{DummySmoother$smooth_fun()}}
\item \href{#method-DummySmoother-clone}{\code{DummySmoother$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Smoother" data-id="export"><a href='../../CorrClim/html/Smoother.html#method-Smoother-export'><code>CorrClim::Smoother$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Smoother" data-id="fit"><a href='../../CorrClim/html/Smoother.html#method-Smoother-fit'><code>CorrClim::Smoother$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Smoother" data-id="fit_smooth"><a href='../../CorrClim/html/Smoother.html#method-Smoother-fit_smooth'><code>CorrClim::Smoother$fit_smooth()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Smoother" data-id="smooth"><a href='../../CorrClim/html/Smoother.html#method-Smoother-smooth'><code>CorrClim::Smoother$smooth()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DummySmoother-new"></a>}}
\if{latex}{\out{\hypertarget{method-DummySmoother-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DummySmoother$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DummySmoother-fit_fun"></a>}}
\if{latex}{\out{\hypertarget{method-DummySmoother-fit_fun}{}}}
\subsection{Method \code{fit_fun()}}{
Fit the Smoother
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DummySmoother$fit_fun(timeseries, y = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} The timeseries to fit on}

\item{\code{y}}{\code{(data.table | TimeseriesDT)} The response timeseries to compare with}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DummySmoother-smooth_fun"></a>}}
\if{latex}{\out{\hypertarget{method-DummySmoother-smooth_fun}{}}}
\subsection{Method \code{smooth_fun()}}{
Apply the Smoother
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DummySmoother$smooth_fun(timeseries)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} The timeseries to smooth}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(TimeseriesDT)}) The timeseries smoothed
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DummySmoother-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DummySmoother-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DummySmoother$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
