% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\docType{class}
\name{ExponentialSmoother}
\alias{ExponentialSmoother}
\title{Exponential smoother}
\description{
This smoother performs an exponential smooth of the timeseries. Can performs a basic exponential smoothing or a exponential smoothing for each timestep of the day, then the comparison is done on the previous day.
}
\section{Super class}{
\code{\link[CorrClim:Smoother]{CorrClim::Smoother}} -> \code{ExponentialSmoother}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{alpha}}{\verb{(float in [0,1])}. Defines as the weight of the smoothing. s(t) = alpha x(t) + (1 - alpha) s(t-1)}

\item{\code{N}}{\code{(Optional[integer])} Default: 20. Only if granularity = 'days'. The number of passed days used to compute the smoothed timeseries}

\item{\code{granularity}}{\code{(string)} \verb{\{'step', 'days'\}}. Default : 'step'. If step, performs a classic EMA step by step, if days, performs a EMA by day.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExponentialSmoother-new}{\code{ExponentialSmoother$new()}}
\item \href{#method-ExponentialSmoother-fit_fun}{\code{ExponentialSmoother$fit_fun()}}
\item \href{#method-ExponentialSmoother-smooth_fun}{\code{ExponentialSmoother$smooth_fun()}}
\item \href{#method-ExponentialSmoother-clone}{\code{ExponentialSmoother$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Smoother" data-id="export"><a href='../../CorrClim/html/Smoother.html#method-Smoother-export'><code>CorrClim::Smoother$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Smoother" data-id="fit"><a href='../../CorrClim/html/Smoother.html#method-Smoother-fit'><code>CorrClim::Smoother$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Smoother" data-id="fit_smooth"><a href='../../CorrClim/html/Smoother.html#method-Smoother-fit_smooth'><code>CorrClim::Smoother$fit_smooth()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Smoother" data-id="smooth"><a href='../../CorrClim/html/Smoother.html#method-Smoother-smooth'><code>CorrClim::Smoother$smooth()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExponentialSmoother-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExponentialSmoother-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExponentialSmoother$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExponentialSmoother-fit_fun"></a>}}
\if{latex}{\out{\hypertarget{method-ExponentialSmoother-fit_fun}{}}}
\subsection{Method \code{fit_fun()}}{
Fit the Smoother
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExponentialSmoother$fit_fun(timeseries, y = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} The timeseries to fit on}

\item{\code{y}}{\code{(data.table | TimeseriesDT)} The response timeseries to compare with}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExponentialSmoother-smooth_fun"></a>}}
\if{latex}{\out{\hypertarget{method-ExponentialSmoother-smooth_fun}{}}}
\subsection{Method \code{smooth_fun()}}{
Apply the Smoother
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExponentialSmoother$smooth_fun(timeseries)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} The timeseries to smooth}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(TimeseriesDT)}) The timeseries smoothed
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExponentialSmoother-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExponentialSmoother-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExponentialSmoother$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
