% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\docType{class}
\name{Formula}
\alias{Formula}
\title{Formula}
\description{
An R6 class which stands as the formula structure of the library
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{formula}}{\code{(str | formula)} The formula of the model. exemple : y ~ temperature + snow}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Formula-new}{\code{Formula$new()}}
\item \href{#method-Formula-get_formula_str}{\code{Formula$get_formula_str()}}
\item \href{#method-Formula-get_formula}{\code{Formula$get_formula()}}
\item \href{#method-Formula-get_formula_base}{\code{Formula$get_formula_base()}}
\item \href{#method-Formula-get_explanatory_variables_formula_base}{\code{Formula$get_explanatory_variables_formula_base()}}
\item \href{#method-Formula-get_all_variables_formula_base}{\code{Formula$get_all_variables_formula_base()}}
\item \href{#method-Formula-get_all_variables}{\code{Formula$get_all_variables()}}
\item \href{#method-Formula-shift_formula}{\code{Formula$shift_formula()}}
\item \href{#method-Formula-get_explanatory_variables}{\code{Formula$get_explanatory_variables()}}
\item \href{#method-Formula-clone}{\code{Formula$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Formula-new"></a>}}
\if{latex}{\out{\hypertarget{method-Formula-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Formula$new(f)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Formula-get_formula_str"></a>}}
\if{latex}{\out{\hypertarget{method-Formula-get_formula_str}{}}}
\subsection{Method \code{get_formula_str()}}{
Get the formula formatted as a character
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Formula$get_formula_str()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{(string)} The formula as a character
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Formula-get_formula"></a>}}
\if{latex}{\out{\hypertarget{method-Formula-get_formula}{}}}
\subsection{Method \code{get_formula()}}{
Get the formula formatted
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Formula$get_formula()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Formula-get_formula_base"></a>}}
\if{latex}{\out{\hypertarget{method-Formula-get_formula_base}{}}}
\subsection{Method \code{get_formula_base()}}{
Get the formula formatted before shift_formula method (if applied)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Formula$get_formula_base()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Formula-get_explanatory_variables_formula_base"></a>}}
\if{latex}{\out{\hypertarget{method-Formula-get_explanatory_variables_formula_base}{}}}
\subsection{Method \code{get_explanatory_variables_formula_base()}}{
Get the explanatory variables from the base formula (before shift method)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Formula$get_explanatory_variables_formula_base()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Formula-get_all_variables_formula_base"></a>}}
\if{latex}{\out{\hypertarget{method-Formula-get_all_variables_formula_base}{}}}
\subsection{Method \code{get_all_variables_formula_base()}}{
Get the all variables from the base formula (before shift method)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Formula$get_all_variables_formula_base()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Formula-get_all_variables"></a>}}
\if{latex}{\out{\hypertarget{method-Formula-get_all_variables}{}}}
\subsection{Method \code{get_all_variables()}}{
Get all the variables (response and explanatory)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Formula$get_all_variables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Formula-shift_formula"></a>}}
\if{latex}{\out{\hypertarget{method-Formula-shift_formula}{}}}
\subsection{Method \code{shift_formula()}}{
Set the formula to add a character to each variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Formula$shift_formula()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Formula-get_explanatory_variables"></a>}}
\if{latex}{\out{\hypertarget{method-Formula-get_explanatory_variables}{}}}
\subsection{Method \code{get_explanatory_variables()}}{
Get explanatory variables from formula
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Formula$get_explanatory_variables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The explanatory variables from the formula
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Formula-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Formula-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Formula$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
