% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.R
\docType{class}
\name{GAM}
\alias{GAM}
\title{GAM (Generative Additive Model)}
\description{
The Generative Additive Model (GAM) for climate correction.
}
\section{Super class}{
\code{\link[CorrClim:TimeseriesModel]{CorrClim::TimeseriesModel}} -> \code{GAM}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{formula}}{\code{(formula)} the formula object representing the fit wanted (e.g y ~ temperature)}

\item{\code{by_instant}}{\code{(boolean)}. If you want to perform a fit by instant of specific granularity}

\item{\code{granularity}}{\code{(Optional[string])} \verb{\{'day', 'month'\}}. The granularity used to compute instant.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{formula}}{\code{(formula)} the formula object representing the fit wanted (e.g y ~ temperature)}

\item{\code{by_instant}}{\code{(boolean)}. If you want to perform a fit by instant of specific granularity}

\item{\code{granularity}}{\code{(Optional[string])} \verb{\{'day', 'month'\}}. The granularity used to compute instant.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GAM-new}{\code{GAM$new()}}
\item \href{#method-GAM-print}{\code{GAM$print()}}
\item \href{#method-GAM-fit_fun}{\code{GAM$fit_fun()}}
\item \href{#method-GAM-predict_fun}{\code{GAM$predict_fun()}}
\item \href{#method-GAM-clone}{\code{GAM$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="check_timeseries"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-check_timeseries'><code>CorrClim::TimeseriesModel$check_timeseries()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="cv_predict"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-cv_predict'><code>CorrClim::TimeseriesModel$cv_predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="export"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-export'><code>CorrClim::TimeseriesModel$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="fit"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-fit'><code>CorrClim::TimeseriesModel$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_formula"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_formula'><code>CorrClim::TimeseriesModel$get_formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_model"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_model'><code>CorrClim::TimeseriesModel$get_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_smoothers"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_smoothers'><code>CorrClim::TimeseriesModel$get_smoothers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="predict"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-predict'><code>CorrClim::TimeseriesModel$predict()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAM-new"></a>}}
\if{latex}{\out{\hypertarget{method-GAM-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAM$new(
  formula = y ~ s(temperature) + s(posan) + jour_semaine + jour_ferie + ponts,
  by_instant = TRUE,
  granularity = "day",
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAM-print"></a>}}
\if{latex}{\out{\hypertarget{method-GAM-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAM$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAM-fit_fun"></a>}}
\if{latex}{\out{\hypertarget{method-GAM-fit_fun}{}}}
\subsection{Method \code{fit_fun()}}{
Fit function of the model itself
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAM$fit_fun(model, X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{\code{(Any)} The model to fit}

\item{\code{X}}{\code{(data.table | TimeseriesDT)} Timeseries data to fit. Should contains all variables in the formula}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAM-predict_fun"></a>}}
\if{latex}{\out{\hypertarget{method-GAM-predict_fun}{}}}
\subsection{Method \code{predict_fun()}}{
Predict function of the model itself
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAM$predict_fun(model, X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{\code{(Any)} The model trained}

\item{\code{X}}{\code{(data.table | TimeseriesDT)} Timeseries data to predict.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(vector)} The output of the prediction of the model on X.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GAM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
