% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.R
\docType{class}
\name{GamStd}
\alias{GamStd}
\title{GamStd (Generative Additive Model for Standard Deviation estimation)}
\description{
The Generative Additive Model (GAM) for timeseries conditional standard deviation estimation.
}
\section{Super classes}{
\code{\link[CorrClim:TimeseriesModel]{CorrClim::TimeseriesModel}} -> \code{\link[CorrClim:TimeseriesStdModel]{CorrClim::TimeseriesStdModel}} -> \code{GamStd}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{formula}}{\code{(formula)} the formula object representing the fit wanted (e.g y ~ temperature)}

\item{\code{by_instant}}{\code{(boolean)}. If you want to perform a fit by instant of specific granularity}

\item{\code{granularity}}{\code{(Optional[string])} \verb{\{'day', 'month'\}}. The granularity used to compute instant.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{formula}}{\code{(formula)} the formula object representing the fit wanted (e.g y ~ temperature)}

\item{\code{by_instant}}{\code{(boolean)}. If you want to perform a fit by instant of specific granularity}

\item{\code{granularity}}{\code{(Optional[string])} \verb{\{'day', 'month'\}}. The granularity used to compute instant.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GamStd-new}{\code{GamStd$new()}}
\item \href{#method-GamStd-print}{\code{GamStd$print()}}
\item \href{#method-GamStd-fit_fun}{\code{GamStd$fit_fun()}}
\item \href{#method-GamStd-predict_fun}{\code{GamStd$predict_fun()}}
\item \href{#method-GamStd-clone}{\code{GamStd$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="check_timeseries"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-check_timeseries'><code>CorrClim::TimeseriesModel$check_timeseries()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="cv_predict"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-cv_predict'><code>CorrClim::TimeseriesModel$cv_predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="export"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-export'><code>CorrClim::TimeseriesModel$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_formula"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_formula'><code>CorrClim::TimeseriesModel$get_formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_model"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_model'><code>CorrClim::TimeseriesModel$get_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_smoothers"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_smoothers'><code>CorrClim::TimeseriesModel$get_smoothers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesStdModel" data-id="fit"><a href='../../CorrClim/html/TimeseriesStdModel.html#method-TimeseriesStdModel-fit'><code>CorrClim::TimeseriesStdModel$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesStdModel" data-id="predict"><a href='../../CorrClim/html/TimeseriesStdModel.html#method-TimeseriesStdModel-predict'><code>CorrClim::TimeseriesStdModel$predict()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GamStd-new"></a>}}
\if{latex}{\out{\hypertarget{method-GamStd-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamStd$new(
  formula = y ~ s(temperature) + s(posan) + jour_semaine + jour_ferie + ponts,
  by_instant = FALSE,
  granularity = "day",
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GamStd-print"></a>}}
\if{latex}{\out{\hypertarget{method-GamStd-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamStd$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GamStd-fit_fun"></a>}}
\if{latex}{\out{\hypertarget{method-GamStd-fit_fun}{}}}
\subsection{Method \code{fit_fun()}}{
Fit function of the model itself
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamStd$fit_fun(model, X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{\code{(Any)} The model to fit}

\item{\code{X}}{\code{(data.table | TimeseriesDT)} Timeseries data to fit. Should contains all variables in the formula}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GamStd-predict_fun"></a>}}
\if{latex}{\out{\hypertarget{method-GamStd-predict_fun}{}}}
\subsection{Method \code{predict_fun()}}{
Predict function of the model itself
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamStd$predict_fun(model, X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{\code{(Any)}}

\item{\code{X}}{\code{(data.table | TimeseriesDT)} Timeseries data to predict.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(list)} The output of the prediction of the model on X. Returns the timeseries predicted (then delta = False) or the delta directly (timeseries is then the initial timeseries).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GamStd-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GamStd-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamStd$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
