% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_delta.R
\docType{class}
\name{GradDelta}
\alias{GradDelta}
\title{GradDelta}
\description{
This class represents a model (GradDelta) for predicting values based on gradients, and linear models. Performs a linear model by instant.
}
\section{Super class}{
\code{\link[CorrClim:TimeseriesModel]{CorrClim::TimeseriesModel}} -> \code{GradDelta}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{formula}}{\code{(formula)} the formula object representing the fit wanted (e.g y ~ temperature)}

\item{\code{N_min}}{\code{(integer)}. Default: 30. Minimum sample allowed to perform the linear regression.}

\item{\code{weights}}{\verb{(float vector)} Weights to pass in the linear regression model. Default is null.}

\item{\code{lm}}{\verb{(MASS linear model)}. Default: 'robust'. Linear model. Choose between robust, least squares and ridge.}

\item{\code{n_shift}}{\code{(integer)} Default: 168. The number of steps in hour to shift your timeseries}

\item{\code{granularity}}{\code{(Optional[string])} \verb{\{'day', 'month'\}}. Default : 'day'. The granularity used to compute instant.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{formula}}{\code{(formula)} the formula object representing the fit wanted (e.g y ~ temperature)}

\item{\code{granularity}}{\code{(Optional[string])} \verb{\{'day', 'month'\}}. Default : 'day'. The granularity used to compute instant.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GradDelta-new}{\code{GradDelta$new()}}
\item \href{#method-GradDelta-print}{\code{GradDelta$print()}}
\item \href{#method-GradDelta-fit_fun}{\code{GradDelta$fit_fun()}}
\item \href{#method-GradDelta-predict_fun}{\code{GradDelta$predict_fun()}}
\item \href{#method-GradDelta-get_gradients}{\code{GradDelta$get_gradients()}}
\item \href{#method-GradDelta-clone}{\code{GradDelta$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="check_timeseries"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-check_timeseries'><code>CorrClim::TimeseriesModel$check_timeseries()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="cv_predict"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-cv_predict'><code>CorrClim::TimeseriesModel$cv_predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="export"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-export'><code>CorrClim::TimeseriesModel$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="fit"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-fit'><code>CorrClim::TimeseriesModel$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_formula"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_formula'><code>CorrClim::TimeseriesModel$get_formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_model"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_model'><code>CorrClim::TimeseriesModel$get_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_smoothers"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_smoothers'><code>CorrClim::TimeseriesModel$get_smoothers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="predict"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-predict'><code>CorrClim::TimeseriesModel$predict()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GradDelta-new"></a>}}
\if{latex}{\out{\hypertarget{method-GradDelta-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GradDelta$new(
  formula = y ~ temperature,
  by_instant = FALSE,
  granularity = "day",
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GradDelta-print"></a>}}
\if{latex}{\out{\hypertarget{method-GradDelta-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GradDelta$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GradDelta-fit_fun"></a>}}
\if{latex}{\out{\hypertarget{method-GradDelta-fit_fun}{}}}
\subsection{Method \code{fit_fun()}}{
Fit function of the model itself
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GradDelta$fit_fun(model, X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{\code{(Any)} The model to fit}

\item{\code{X}}{\code{(data.table | TimeseriesDT)} Timeseries data to fit. Should contains all variables in the formula}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GradDelta-predict_fun"></a>}}
\if{latex}{\out{\hypertarget{method-GradDelta-predict_fun}{}}}
\subsection{Method \code{predict_fun()}}{
Predict function of the model itself
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GradDelta$predict_fun(model, X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{\code{(Any)} The model trained}

\item{\code{X}}{\code{(data.table | TimeseriesDT)} Timeseries data to predict.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(vector)} The output of the prediction of the model on X.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GradDelta-get_gradients"></a>}}
\if{latex}{\out{\hypertarget{method-GradDelta-get_gradients}{}}}
\subsection{Method \code{get_gradients()}}{
Get gradients from the GradDelta model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GradDelta$get_gradients()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GradDelta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GradDelta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GradDelta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
