% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\docType{class}
\name{GridSearchSmoother}
\alias{GridSearchSmoother}
\title{GridSearch Exponential Smoother}
\description{
Perform a Grid search on a Smoother to find best params and smooth using best smoother.
}
\section{Super class}{
\code{\link[CorrClim:Smoother]{CorrClim::Smoother}} -> \code{GridSearchSmoother}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{grid}}{\code{(list)} The list of the parameters and the sequences associated to test. Example : list(alpha = seq(0.5, 1, 0.01))}

\item{\code{score}}{\code{(callable)} Default : correlation_score . A callable to compute the score to maximize. Should return a float and take a data.table with two columns in input.}

\item{\code{smoother}}{\code{(Smoother)} A smoother object not instantiated to perform the gridsearch on.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GridSearchSmoother-new}{\code{GridSearchSmoother$new()}}
\item \href{#method-GridSearchSmoother-fit_fun}{\code{GridSearchSmoother$fit_fun()}}
\item \href{#method-GridSearchSmoother-smooth_fun}{\code{GridSearchSmoother$smooth_fun()}}
\item \href{#method-GridSearchSmoother-get_best_params}{\code{GridSearchSmoother$get_best_params()}}
\item \href{#method-GridSearchSmoother-get_best_smoother}{\code{GridSearchSmoother$get_best_smoother()}}
\item \href{#method-GridSearchSmoother-clone}{\code{GridSearchSmoother$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Smoother" data-id="export"><a href='../../CorrClim/html/Smoother.html#method-Smoother-export'><code>CorrClim::Smoother$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Smoother" data-id="fit"><a href='../../CorrClim/html/Smoother.html#method-Smoother-fit'><code>CorrClim::Smoother$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Smoother" data-id="fit_smooth"><a href='../../CorrClim/html/Smoother.html#method-Smoother-fit_smooth'><code>CorrClim::Smoother$fit_smooth()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Smoother" data-id="smooth"><a href='../../CorrClim/html/Smoother.html#method-Smoother-smooth'><code>CorrClim::Smoother$smooth()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridSearchSmoother-new"></a>}}
\if{latex}{\out{\hypertarget{method-GridSearchSmoother-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridSearchSmoother$new(grid, smoother, score = correlation_score, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridSearchSmoother-fit_fun"></a>}}
\if{latex}{\out{\hypertarget{method-GridSearchSmoother-fit_fun}{}}}
\subsection{Method \code{fit_fun()}}{
Fit the smoother
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridSearchSmoother$fit_fun(timeseries, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} The timeseries to fit on}

\item{\code{y}}{\code{(data.table | TimeseriesDT)} The response timeseries to compare with}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridSearchSmoother-smooth_fun"></a>}}
\if{latex}{\out{\hypertarget{method-GridSearchSmoother-smooth_fun}{}}}
\subsection{Method \code{smooth_fun()}}{
Apply the Smoother
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridSearchSmoother$smooth_fun(timeseries)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} The timeseries to smooth}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(TimeseriesDT)}) The timeseries smoothed
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridSearchSmoother-get_best_params"></a>}}
\if{latex}{\out{\hypertarget{method-GridSearchSmoother-get_best_params}{}}}
\subsection{Method \code{get_best_params()}}{
Get the best parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridSearchSmoother$get_best_params()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{(float)} Optimal alpha
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridSearchSmoother-get_best_smoother"></a>}}
\if{latex}{\out{\hypertarget{method-GridSearchSmoother-get_best_smoother}{}}}
\subsection{Method \code{get_best_smoother()}}{
Get the best smoother
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridSearchSmoother$get_best_smoother()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{(Smoother)} Best smoother
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridSearchSmoother-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GridSearchSmoother-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridSearchSmoother$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
