% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{class}
\name{Metrics}
\alias{Metrics}
\title{Metrics}
\description{
Metrics model-agnostic
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Metrics-new}{\code{Metrics$new()}}
\item \href{#method-Metrics-plot_timeseries}{\code{Metrics$plot_timeseries()}}
\item \href{#method-Metrics-RMSE}{\code{Metrics$RMSE()}}
\item \href{#method-Metrics-MAPE}{\code{Metrics$MAPE()}}
\item \href{#method-Metrics-BIAS}{\code{Metrics$BIAS()}}
\item \href{#method-Metrics-plot_bar_volume_correction}{\code{Metrics$plot_bar_volume_correction()}}
\item \href{#method-Metrics-plot_aggregated_cc_evol_by_category}{\code{Metrics$plot_aggregated_cc_evol_by_category()}}
\item \href{#method-Metrics-plot_scatter_timeseries}{\code{Metrics$plot_scatter_timeseries()}}
\item \href{#method-Metrics-clone}{\code{Metrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metrics-new"></a>}}
\if{latex}{\out{\hypertarget{method-Metrics-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metrics$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metrics-plot_timeseries"></a>}}
\if{latex}{\out{\hypertarget{method-Metrics-plot_timeseries}{}}}
\subsection{Method \code{plot_timeseries()}}{
Plot timeseries before and after the climate correction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metrics$plot_timeseries(
  timeseries,
  prediction,
  ts_to_compare_with = NULL,
  variables_to_compare = NULL,
  granularity = "day",
  fun = mean,
  ylabel = "Timeseries",
  ylabel_to_compare = NULL,
  title = "Timeseries comparison with and without climate correction",
  legend = c("Timeseries input", "Timeseries climate corrected"),
  legend_difference = "Correction climatique",
  width = 1000,
  height = 700
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} timeseries with measures}

\item{\code{prediction}}{\code{(data.table | TimeseriesDT)} timeseries with forecasts/climate corrections}

\item{\code{ts_to_compare_with}}{\code{(data.table | TimeseriesDT)} The timeseries to compare with. Builds a subplots if feeded.}

\item{\code{variables_to_compare}}{(vector of string) The variables in the ts_to_compare_with data to plot.}

\item{\code{granularity}}{\code{(string)} The granularity to aggregate on. Choose between hour, days, week, month, year}

\item{\code{fun}}{\code{(callable)} The aggregation function to apply.}

\item{\code{ylabel}}{\code{(string)} The label for y axis.}

\item{\code{ylabel_to_compare}}{\code{(string)} The label for the y axis timeseries to compare}

\item{\code{title}}{\code{(string)} Title of the graph}

\item{\code{legend}}{(vector of string) The legend for timeseries and prediction args}

\item{\code{legend_difference}}{\code{(string)} The legend for the difference between timeseries and prediction}

\item{\code{width}}{\code{(string)} The width of the graph}

\item{\code{height}}{\code{(integer)} The height of the graph}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{(plotly graph)} The before/after timeseries
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metrics-RMSE"></a>}}
\if{latex}{\out{\hypertarget{method-Metrics-RMSE}{}}}
\subsection{Method \code{RMSE()}}{
Root Mean Squared Error
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metrics$RMSE(
  timeseries,
  prediction,
  plot = TRUE,
  granularity = "day",
  fun = mean,
  width = 1000,
  height = 700
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} timeseries with measures}

\item{\code{prediction}}{\code{(data.table | TimeseriesDT)} timeseries with forecasts/climate corrections}

\item{\code{plot}}{\code{(boolean)} Whether to plot or not the evolution of RSE}

\item{\code{granularity}}{\code{(string)} The granularity to aggregate the timeseries on (used only for plots)}

\item{\code{fun}}{\code{(callable)} An aggregation function for the timeseries (used only for plots)}

\item{\code{width}}{\code{(integer)} The width of the graph}

\item{\code{height}}{\code{(integer)} The height of the graph}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{(float | plotly graph)} RMSE plot or RMSE float
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metrics-MAPE"></a>}}
\if{latex}{\out{\hypertarget{method-Metrics-MAPE}{}}}
\subsection{Method \code{MAPE()}}{
Mean Absolute Pourcentage Error
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metrics$MAPE(
  timeseries,
  prediction,
  plot = TRUE,
  granularity = "day",
  fun = mean,
  width = 1000,
  height = 700
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} timeseries with measures}

\item{\code{prediction}}{\code{(data.table | TimeseriesDT)} timeseries with forecasts/climate corrections}

\item{\code{plot}}{\code{(boolean)} Whether to plot or not the evolution of APE}

\item{\code{granularity}}{\code{(string)} The granularity to aggregate the timeseries on (used only for plots)}

\item{\code{fun}}{\code{(callable)} An aggregation function for the timeseries (used only for plots)}

\item{\code{width}}{\code{(integer)} The width of the graph}

\item{\code{height}}{\code{(integer)} The height of the graph}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{(float | plotly graph)} MAPE plot or MAPE float
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metrics-BIAS"></a>}}
\if{latex}{\out{\hypertarget{method-Metrics-BIAS}{}}}
\subsection{Method \code{BIAS()}}{
Bias (error = measure - forecast/cc)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metrics$BIAS(
  timeseries,
  prediction,
  plot = TRUE,
  granularity = "day",
  fun = mean,
  width = 1000,
  height = 700
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} timeseries with measures}

\item{\code{prediction}}{\code{(data.table | TimeseriesDT)} timeseries with forecasts/climate corrections}

\item{\code{plot}}{\code{(boolean)} Whether to plot or not the evolution of APE}

\item{\code{granularity}}{\code{(string)} The granularity to aggregate the timeseries on (used only for plots)}

\item{\code{fun}}{\code{(callable)} An aggregation function for the timeseries (used only for plots)}

\item{\code{width}}{\code{(integer)} The width of the graph}

\item{\code{height}}{\code{(integer)} The height of the graph}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{(float | plotly graph)} BIAS plot or BIAS float
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metrics-plot_bar_volume_correction"></a>}}
\if{latex}{\out{\hypertarget{method-Metrics-plot_bar_volume_correction}{}}}
\subsection{Method \code{plot_bar_volume_correction()}}{
Plot the volume of climate correction aggregated on a specific granularity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metrics$plot_bar_volume_correction(
  timeseries,
  prediction,
  granularity = "month",
  fun = mean,
  title = "Climate correction volume",
  ylabel = "Energy",
  width = 1000,
  height = 700
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} timeseries with measures}

\item{\code{prediction}}{\code{(data.table | TimeseriesDT)} timeseries with forecasts/climate corrections}

\item{\code{granularity}}{\code{(string)} The granularity to aggregate the timeseries on (used only for plots)}

\item{\code{fun}}{\code{(callable)} An aggregation function for the timeseries (used only for plots)}

\item{\code{title}}{\code{(string)} The graph title}

\item{\code{ylabel}}{\code{(string)} The label for y axis.}

\item{\code{width}}{\code{(integer)} The width of the graph}

\item{\code{height}}{\code{(integer)} The height of the graph}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{(plotly graph)} Barplot of the climate correction volume
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metrics-plot_aggregated_cc_evol_by_category"></a>}}
\if{latex}{\out{\hypertarget{method-Metrics-plot_aggregated_cc_evol_by_category}{}}}
\subsection{Method \code{plot_aggregated_cc_evol_by_category()}}{
Plot the evolution through time of climate correction aggregated at a specific granularity, for different categories
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metrics$plot_aggregated_cc_evol_by_category(
  prediction,
  category_varname,
  granularity = "year",
  fun = sum,
  title = "Climate correction evolution by category",
  ylabel = "Energy",
  width = 1000,
  height = 700
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prediction}}{(data.table | TimeseriesDT | list there of) list of timeseries with forecasts/climate corrections (1 element for each category)}

\item{\code{granularity}}{\code{(string)} The granularity to aggregate the timeseries on (used only for plots)}

\item{\code{fun}}{\code{(callable)} An aggregation function for the timeseries (used only for plots)}

\item{\code{title}}{\code{(string)} The graph title}

\item{\code{ylabel}}{\code{(string)} The label for y axis.}

\item{\code{width}}{\code{(integer)} The width of the graph}

\item{\code{height}}{\code{(integer)} The height of the graph}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{(plotly graph)} Barplot of the climate correction volume
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metrics-plot_scatter_timeseries"></a>}}
\if{latex}{\out{\hypertarget{method-Metrics-plot_scatter_timeseries}{}}}
\subsection{Method \code{plot_scatter_timeseries()}}{
Plot a scatter of y_ts vs x_ts. Add second_y_ts if you want two scatter group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metrics$plot_scatter_timeseries(
  y_ts,
  x_ts,
  xlabel,
  ylabel,
  title,
  second_y_ts = NULL,
  legend = c("First Timeseries", "Second Timeseries"),
  opacity = 0.7,
  granularity = "day",
  width = 1000,
  height = 700,
  size = 6,
  fun = mean
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_ts}}{\code{(data.table | TimeseriesDT)} The timeseries to plot in y axis}

\item{\code{x_ts}}{\code{(data.table | TimeseriesDT)} The timeseries to plot in x axis}

\item{\code{xlabel}}{\code{(string)} The label for x axis}

\item{\code{ylabel}}{\code{(string)} The label for y axis}

\item{\code{title}}{\code{(string)} The graph title}

\item{\code{second_y_ts}}{\code{(data.table | TimeseriesDT)} The second timeseries to plot in y axis}

\item{\code{granularity}}{\code{(string)} The granularity to aggregate the timeseries}

\item{\code{width}}{\code{(integer)} The width of the graph}

\item{\code{height}}{\code{(integer)} The height of the graph}

\item{\code{size}}{\code{(integer)} Dot size of the scatter plot}

\item{\code{fun}}{\code{(callable)} An aggregation function for the timeseries}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{(float | plotly graph)} Scatter plot for timeseries
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metrics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Metrics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
