% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\docType{class}
\name{MultiSmoother}
\alias{MultiSmoother}
\title{MultiSmoother}
\description{
A MultiSmoother class to encapsulates the role of multiple smoothing terms. Smoothes only temperature variables, and consider one smoother for one temperature variable.
}
\section{Super class}{
\code{\link[CorrClim:Smoother]{CorrClim::Smoother}} -> \code{MultiSmoother}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{smoothers}}{(vector of Smoother object). The smoothers to apply.}

\item{\code{variables}}{(vector of strings). The variables to smooth.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultiSmoother-new}{\code{MultiSmoother$new()}}
\item \href{#method-MultiSmoother-fit_fun}{\code{MultiSmoother$fit_fun()}}
\item \href{#method-MultiSmoother-smooth_fun}{\code{MultiSmoother$smooth_fun()}}
\item \href{#method-MultiSmoother-get_smoothers}{\code{MultiSmoother$get_smoothers()}}
\item \href{#method-MultiSmoother-clone}{\code{MultiSmoother$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Smoother" data-id="export"><a href='../../CorrClim/html/Smoother.html#method-Smoother-export'><code>CorrClim::Smoother$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Smoother" data-id="fit"><a href='../../CorrClim/html/Smoother.html#method-Smoother-fit'><code>CorrClim::Smoother$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Smoother" data-id="fit_smooth"><a href='../../CorrClim/html/Smoother.html#method-Smoother-fit_smooth'><code>CorrClim::Smoother$fit_smooth()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Smoother" data-id="smooth"><a href='../../CorrClim/html/Smoother.html#method-Smoother-smooth'><code>CorrClim::Smoother$smooth()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiSmoother-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultiSmoother-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiSmoother$new(smoothers, variables)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiSmoother-fit_fun"></a>}}
\if{latex}{\out{\hypertarget{method-MultiSmoother-fit_fun}{}}}
\subsection{Method \code{fit_fun()}}{
Apply the Smoother
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiSmoother$fit_fun(timeseries, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} The timeseries to smooth}

\item{\code{y}}{\code{(data.table | TimeseriesDT)} The response timeseries to compare with}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiSmoother-smooth_fun"></a>}}
\if{latex}{\out{\hypertarget{method-MultiSmoother-smooth_fun}{}}}
\subsection{Method \code{smooth_fun()}}{
Apply the Smoother
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiSmoother$smooth_fun(timeseries)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} The timeseries to smooth}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(TimeseriesDT)}) The timeseries smoothed
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiSmoother-get_smoothers"></a>}}
\if{latex}{\out{\hypertarget{method-MultiSmoother-get_smoothers}{}}}
\subsection{Method \code{get_smoothers()}}{
Get smoothers of the MultiSmoother object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiSmoother$get_smoothers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiSmoother-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultiSmoother-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiSmoother$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
