% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\docType{class}
\name{Operator}
\alias{Operator}
\title{Operator}
\description{
Represents the structure of Climate Correction Operator.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Operator-new}{\code{Operator$new()}}
\item \href{#method-Operator-apply}{\code{Operator$apply()}}
\item \href{#method-Operator-clone}{\code{Operator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Operator-new"></a>}}
\if{latex}{\out{\hypertarget{method-Operator-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Operator$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Operator-apply"></a>}}
\if{latex}{\out{\hypertarget{method-Operator-apply}{}}}
\subsection{Method \code{apply()}}{
Apply the Operator on the output to make the climatic correction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Operator$apply(
  timeseries,
  y_pred_observed,
  y_pred_target,
  y_std_observed = NULL,
  y_std_target = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} The timeseries to apply climate correction on.}

\item{\code{y_pred_observed}}{\code{(vector)} The inference made on the observed weather.}

\item{\code{y_pred_target}}{\code{(vector)} The inference made on the target weather}

\item{\code{y_std_observed}}{\code{(vector)} The inference of standard deviation made on the observed weather}

\item{\code{y_std_target}}{\code{(vector)} The inference of standard deviation made on the target weather}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(TimeseriesDT)} The output climate corrected
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Operator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Operator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Operator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
