% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\docType{class}
\name{Operator2Moments}
\alias{Operator2Moments}
\title{Operator2Moments}
\description{
Operator to perform climatic correction over the 2 first distribution moments (expectation and standard deviation)
}
\section{Super class}{
\code{\link[CorrClim:Operator]{CorrClim::Operator}} -> \code{Operator2Moments}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Operator2Moments-new}{\code{Operator2Moments$new()}}
\item \href{#method-Operator2Moments-apply_fun}{\code{Operator2Moments$apply_fun()}}
\item \href{#method-Operator2Moments-clone}{\code{Operator2Moments$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Operator" data-id="apply"><a href='../../CorrClim/html/Operator.html#method-Operator-apply'><code>CorrClim::Operator$apply()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Operator2Moments-new"></a>}}
\if{latex}{\out{\hypertarget{method-Operator2Moments-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Operator2Moments$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Operator2Moments-apply_fun"></a>}}
\if{latex}{\out{\hypertarget{method-Operator2Moments-apply_fun}{}}}
\subsection{Method \code{apply_fun()}}{
Apply the 2-moments Operator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Operator2Moments$apply_fun(
  timeseries,
  y_pred_observed,
  y_pred_target,
  y_std_observed,
  y_std_target
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} The timeseries to apply climate correction on.}

\item{\code{y_pred_observed}}{\code{(vector)} The inference made on the observed weather.}

\item{\code{y_pred_target}}{\code{(vector)} The inference made on the target weather}

\item{\code{y_std_observed}}{\code{(vector)} The inference of standard deviation made on the observed weather}

\item{\code{y_std_target}}{\code{(vector)} The inference of standard deviation made on the target weather}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(TimeseriesDT)} The output climate corrected
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Operator2Moments-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Operator2Moments-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Operator2Moments$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
