% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\docType{class}
\name{OperatorAdditive}
\alias{OperatorAdditive}
\title{OperatorAdditive}
\description{
Additive Operator. Performs an addition of the climate part of the timeseries
}
\section{Super class}{
\code{\link[CorrClim:Operator]{CorrClim::Operator}} -> \code{OperatorAdditive}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OperatorAdditive-new}{\code{OperatorAdditive$new()}}
\item \href{#method-OperatorAdditive-apply_fun}{\code{OperatorAdditive$apply_fun()}}
\item \href{#method-OperatorAdditive-clone}{\code{OperatorAdditive$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Operator" data-id="apply"><a href='../../CorrClim/html/Operator.html#method-Operator-apply'><code>CorrClim::Operator$apply()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperatorAdditive-new"></a>}}
\if{latex}{\out{\hypertarget{method-OperatorAdditive-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperatorAdditive$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperatorAdditive-apply_fun"></a>}}
\if{latex}{\out{\hypertarget{method-OperatorAdditive-apply_fun}{}}}
\subsection{Method \code{apply_fun()}}{
Apply function of the Additive operator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperatorAdditive$apply_fun(timeseries, y_pred_observed, y_pred_target, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} The timeseries to apply climate correction on.}

\item{\code{y_pred_observed}}{\code{(vector)} The inference made on the observed weather.}

\item{\code{y_pred_target}}{\code{(vector)} The inference made on the target weather}

\item{\code{...}}{Additional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(TimeseriesDT)} The output climate corrected
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperatorAdditive-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OperatorAdditive-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperatorAdditive$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
