% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\docType{class}
\name{OperatorMultiplicative}
\alias{OperatorMultiplicative}
\title{OperatorMultiplicative}
\description{
Multiplicative Operator. Apply a share of the climate part of the timeseries
}
\section{Super class}{
\code{\link[CorrClim:Operator]{CorrClim::Operator}} -> \code{OperatorMultiplicative}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OperatorMultiplicative-new}{\code{OperatorMultiplicative$new()}}
\item \href{#method-OperatorMultiplicative-apply_fun}{\code{OperatorMultiplicative$apply_fun()}}
\item \href{#method-OperatorMultiplicative-clone}{\code{OperatorMultiplicative$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Operator" data-id="apply"><a href='../../CorrClim/html/Operator.html#method-Operator-apply'><code>CorrClim::Operator$apply()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperatorMultiplicative-new"></a>}}
\if{latex}{\out{\hypertarget{method-OperatorMultiplicative-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperatorMultiplicative$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperatorMultiplicative-apply_fun"></a>}}
\if{latex}{\out{\hypertarget{method-OperatorMultiplicative-apply_fun}{}}}
\subsection{Method \code{apply_fun()}}{
Apply the Multiplicative Operator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperatorMultiplicative$apply_fun(
  timeseries,
  y_pred_observed,
  y_pred_target,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} The timeseries to apply climate correction on.}

\item{\code{y_pred_observed}}{\code{(vector)} The inference made on the observed weather.}

\item{\code{y_pred_target}}{\code{(vector)} The inference made on the target weather}

\item{\code{...}}{Additional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(TimeseriesDT)} The output climate corrected
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperatorMultiplicative-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OperatorMultiplicative-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperatorMultiplicative$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
