% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\docType{class}
\name{OperatorTarget}
\alias{OperatorTarget}
\title{OperatorTarget}
\description{
Operator returning model estimation at target weather as a climatic correction
}
\section{Super class}{
\code{\link[CorrClim:Operator]{CorrClim::Operator}} -> \code{OperatorTarget}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OperatorTarget-new}{\code{OperatorTarget$new()}}
\item \href{#method-OperatorTarget-apply_fun}{\code{OperatorTarget$apply_fun()}}
\item \href{#method-OperatorTarget-clone}{\code{OperatorTarget$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="Operator" data-id="apply"><a href='../../CorrClim/html/Operator.html#method-Operator-apply'><code>CorrClim::Operator$apply()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperatorTarget-new"></a>}}
\if{latex}{\out{\hypertarget{method-OperatorTarget-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperatorTarget$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperatorTarget-apply_fun"></a>}}
\if{latex}{\out{\hypertarget{method-OperatorTarget-apply_fun}{}}}
\subsection{Method \code{apply_fun()}}{
Apply function of the Target operator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperatorTarget$apply_fun(y_pred_target, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_pred_target}}{\code{(data.table | TimeseriesDT)} The inference made on the target weather}

\item{\code{...}}{Additional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(TimeseriesDT)} The output climate corrected
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperatorTarget-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OperatorTarget-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperatorTarget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
