% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\docType{class}
\name{Prophet}
\alias{Prophet}
\title{Prophet Model}
\description{
This class represents a model (Prophet) for time series forecasting using the Prophet package.
}
\section{Super class}{
\code{\link[CorrClim:TimeseriesModel]{CorrClim::TimeseriesModel}} -> \code{Prophet}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{formula}}{\code{(formula)} the formula object representing the fit wanted (e.g y ~ temperature)}

\item{\code{yearly_seasonality}}{Boolean, default TRUE. Indicates whether to include yearly seasonality.}

\item{\code{weekly_seasonality}}{Boolean, default TRUE. Indicates whether to include weekly seasonality.}

\item{\code{daily_seasonality}}{Boolean, default TRUE. Indicates whether to include daily seasonality.}

\item{\code{interval_width}}{Numeric, default 0. Indicates the uncertainty interval width.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{formula}}{\code{(formula)} the formula object representing the fit wanted (e.g y ~ temperature)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Prophet-new}{\code{Prophet$new()}}
\item \href{#method-Prophet-print}{\code{Prophet$print()}}
\item \href{#method-Prophet-fit_fun}{\code{Prophet$fit_fun()}}
\item \href{#method-Prophet-predict_fun}{\code{Prophet$predict_fun()}}
\item \href{#method-Prophet-clone}{\code{Prophet$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="check_timeseries"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-check_timeseries'><code>CorrClim::TimeseriesModel$check_timeseries()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="cv_predict"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-cv_predict'><code>CorrClim::TimeseriesModel$cv_predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="export"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-export'><code>CorrClim::TimeseriesModel$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="fit"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-fit'><code>CorrClim::TimeseriesModel$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_formula"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_formula'><code>CorrClim::TimeseriesModel$get_formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_model"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_model'><code>CorrClim::TimeseriesModel$get_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_smoothers"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_smoothers'><code>CorrClim::TimeseriesModel$get_smoothers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="predict"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-predict'><code>CorrClim::TimeseriesModel$predict()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Prophet-new"></a>}}
\if{latex}{\out{\hypertarget{method-Prophet-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Prophet$new(formula = y ~ temperature, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Prophet-print"></a>}}
\if{latex}{\out{\hypertarget{method-Prophet-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Prophet$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Prophet-fit_fun"></a>}}
\if{latex}{\out{\hypertarget{method-Prophet-fit_fun}{}}}
\subsection{Method \code{fit_fun()}}{
Fit function of the model itself
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Prophet$fit_fun(model, X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{\code{(Any)}}

\item{\code{X}}{\code{(data.table | TimeseriesDT)} Timeseries data to fit. Should contains all variables in the formula}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Prophet-predict_fun"></a>}}
\if{latex}{\out{\hypertarget{method-Prophet-predict_fun}{}}}
\subsection{Method \code{predict_fun()}}{
Predict function of the model itself
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Prophet$predict_fun(model, X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{\code{(Any)} The model trained}

\item{\code{X}}{\code{(data.table | TimeseriesDT)} Timeseries data to predict.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(vector)} The output of the prediction of the model on X.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Prophet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Prophet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Prophet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
