% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\docType{class}
\name{Smoother}
\alias{Smoother}
\title{Smoother}
\description{
Base class for any smoother object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Smoother-new}{\code{Smoother$new()}}
\item \href{#method-Smoother-fit}{\code{Smoother$fit()}}
\item \href{#method-Smoother-smooth}{\code{Smoother$smooth()}}
\item \href{#method-Smoother-fit_smooth}{\code{Smoother$fit_smooth()}}
\item \href{#method-Smoother-export}{\code{Smoother$export()}}
\item \href{#method-Smoother-clone}{\code{Smoother$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Smoother-new"></a>}}
\if{latex}{\out{\hypertarget{method-Smoother-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Smoother$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Smoother-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Smoother-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the Smoother
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Smoother$fit(timeseries, y = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} The timeseries to fit on}

\item{\code{y}}{\code{(data.table | TimeseriesDT)} The response timeseries to compare with}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Smoother-smooth"></a>}}
\if{latex}{\out{\hypertarget{method-Smoother-smooth}{}}}
\subsection{Method \code{smooth()}}{
Apply the Smoother
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Smoother$smooth(timeseries)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} The timeseries to smooth}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(TimeseriesDT)} The timeseries smoothed
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Smoother-fit_smooth"></a>}}
\if{latex}{\out{\hypertarget{method-Smoother-fit_smooth}{}}}
\subsection{Method \code{fit_smooth()}}{
Fit and apply the smoother. Calls the fit and smooth method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Smoother$fit_smooth(timeseries, y = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table | TimeseriesDT)} The timeseries to smooth}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(TimeseriesDT)} The timeseries smoothed
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Smoother-export"></a>}}
\if{latex}{\out{\hypertarget{method-Smoother-export}{}}}
\subsection{Method \code{export()}}{
Export the Smoother in RDS.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Smoother$export(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{\code{(string)} Path file to write the rds file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Smoother-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Smoother-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Smoother$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
