% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_dt.R
\docType{class}
\name{TimeseriesDT}
\alias{TimeseriesDT}
\title{TimeseriesDT}
\description{
An R6 class which stands as the timeseries structure described in the library
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{timeseries}}{\code{(data.table)} A data.table which contains a time column and N value columns.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeseriesDT-new}{\code{TimeseriesDT$new()}}
\item \href{#method-TimeseriesDT-remove_na}{\code{TimeseriesDT$remove_na()}}
\item \href{#method-TimeseriesDT-set_timeseries}{\code{TimeseriesDT$set_timeseries()}}
\item \href{#method-TimeseriesDT-get_timeseries}{\code{TimeseriesDT$get_timeseries()}}
\item \href{#method-TimeseriesDT-check_pattern_in_column}{\code{TimeseriesDT$check_pattern_in_column()}}
\item \href{#method-TimeseriesDT-set_format_date}{\code{TimeseriesDT$set_format_date()}}
\item \href{#method-TimeseriesDT-set_timezone}{\code{TimeseriesDT$set_timezone()}}
\item \href{#method-TimeseriesDT-get_variables_from_pattern}{\code{TimeseriesDT$get_variables_from_pattern()}}
\item \href{#method-TimeseriesDT-get_variables_name}{\code{TimeseriesDT$get_variables_name()}}
\item \href{#method-TimeseriesDT-get}{\code{TimeseriesDT$get()}}
\item \href{#method-TimeseriesDT-add_calendar}{\code{TimeseriesDT$add_calendar()}}
\item \href{#method-TimeseriesDT-shift}{\code{TimeseriesDT$shift()}}
\item \href{#method-TimeseriesDT-nrows}{\code{TimeseriesDT$nrows()}}
\item \href{#method-TimeseriesDT-compute_instant}{\code{TimeseriesDT$compute_instant()}}
\item \href{#method-TimeseriesDT-add_suffix}{\code{TimeseriesDT$add_suffix()}}
\item \href{#method-TimeseriesDT-sort}{\code{TimeseriesDT$sort()}}
\item \href{#method-TimeseriesDT-compute_period_start}{\code{TimeseriesDT$compute_period_start()}}
\item \href{#method-TimeseriesDT-aggregate}{\code{TimeseriesDT$aggregate()}}
\item \href{#method-TimeseriesDT-groupby}{\code{TimeseriesDT$groupby()}}
\item \href{#method-TimeseriesDT-select}{\code{TimeseriesDT$select()}}
\item \href{#method-TimeseriesDT-remove_duplicated}{\code{TimeseriesDT$remove_duplicated()}}
\item \href{#method-TimeseriesDT-assign}{\code{TimeseriesDT$assign()}}
\item \href{#method-TimeseriesDT-merge}{\code{TimeseriesDT$merge()}}
\item \href{#method-TimeseriesDT-remove_variables}{\code{TimeseriesDT$remove_variables()}}
\item \href{#method-TimeseriesDT-align}{\code{TimeseriesDT$align()}}
\item \href{#method-TimeseriesDT-compute_degre_days}{\code{TimeseriesDT$compute_degre_days()}}
\item \href{#method-TimeseriesDT-get_granularity}{\code{TimeseriesDT$get_granularity()}}
\item \href{#method-TimeseriesDT-rename}{\code{TimeseriesDT$rename()}}
\item \href{#method-TimeseriesDT-filter_dataset}{\code{TimeseriesDT$filter_dataset()}}
\item \href{#method-TimeseriesDT-export}{\code{TimeseriesDT$export()}}
\item \href{#method-TimeseriesDT-clone}{\code{TimeseriesDT$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-new"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$new(
  timeseries,
  is_output = FALSE,
  format_date = "\%Y-\%m-\%d \%H:\%M:\%S",
  timezone = "UTC"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-remove_na"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-remove_na}{}}}
\subsection{Method \code{remove_na()}}{
Remove NA from timeseries data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$remove_na(inplace = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inplace}}{\code{(boolean)} If TRUE or FALSE a new instance is wanted or modifiying existing one is sufficient.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-set_timeseries"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-set_timeseries}{}}}
\subsection{Method \code{set_timeseries()}}{
Set the timeseries of a TimeseriesDT existing object. Use with caution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$set_timeseries(timeseries)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeseries}}{\code{(data.table)}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-get_timeseries"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-get_timeseries}{}}}
\subsection{Method \code{get_timeseries()}}{
Get the timeseries data.table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$get_timeseries()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{(data.table)} TimeseriesDT data.table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-check_pattern_in_column"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-check_pattern_in_column}{}}}
\subsection{Method \code{check_pattern_in_column()}}{
Look for patterns in a character
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$check_pattern_in_column(patterns, column)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{patterns}}{\verb{(string vector)}}

\item{\code{column}}{\verb{(string vector)}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(boolean vector) A vector asserting is yes or no the pattern is contained in the column arg
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-set_format_date"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-set_format_date}{}}}
\subsection{Method \code{set_format_date()}}{
Set a new format for date
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$set_format_date(format_date = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{format_date}}{\code{(string)} A valid format for dates}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-set_timezone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-set_timezone}{}}}
\subsection{Method \code{set_timezone()}}{
Set a new timezone
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$set_timezone(timezone)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-get_variables_from_pattern"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-get_variables_from_pattern}{}}}
\subsection{Method \code{get_variables_from_pattern()}}{
Get variables from a Timeseries matching patterns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$get_variables_from_pattern(patterns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{patterns}}{\verb{(string vector)}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-get_variables_name"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-get_variables_name}{}}}
\subsection{Method \code{get_variables_name()}}{
Get variables from the dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$get_variables_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-get"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-get}{}}}
\subsection{Method \code{get()}}{
Get variable from timeseries. Similar as dt$var or dt[\link{var}]
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$get(var)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-add_calendar"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-add_calendar}{}}}
\subsection{Method \code{add_calendar()}}{
Generate calendar from a timeseries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$add_calendar(
  variables = c("date", "Annee", "Mois", "Jour", "Heure", "Minute", "Posan", "Tendance",
    "JourSemaine", "JourFerie", "Ponts", "isoSemaine", "isoAnnee"),
  inplace = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variables}}{\verb{(string vector)}variables to add to the calendar.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The calendar generated
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-shift"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-shift}{}}}
\subsection{Method \code{shift()}}{
Add shifted timeseries to existing timeseries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$shift(variables, n = 168, inplace = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variables}}{\verb{(string vector)}variables names to shift}

\item{\code{n}}{\code{(integer)} number of steps to shift on}

\item{\code{inplace}}{\code{(boolean)} If TRUE or FALSE you want to add to existing timeseries or return a new object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-nrows"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-nrows}{}}}
\subsection{Method \code{nrows()}}{
Get the row number of the timeseries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$nrows()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-compute_instant"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-compute_instant}{}}}
\subsection{Method \code{compute_instant()}}{
Add an instant column to existing timeseries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$compute_instant(granularity = "day", inplace = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inplace}}{\code{(boolean)} If TRUE or FALSE you want to add to existing timeseries or return a new object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-add_suffix"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-add_suffix}{}}}
\subsection{Method \code{add_suffix()}}{
Add a suffix to explanatory variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$add_suffix(variables, suffix, inplace = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variables}}{\verb{(string vector)} The variables to add suffix on}

\item{\code{suffix}}{\code{(string)} The suffix you want to add}

\item{\code{inplace}}{\code{(boolean)} If TRUE or FALSE a new instance is wanted or modifiying existing one is sufficient.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-sort"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-sort}{}}}
\subsection{Method \code{sort()}}{
Sort timeseries accordin to a variable, ascending mode.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$sort(variable, inplace = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{\code{(string)} The variable to use to order (ascending mode only) the timeseries}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-compute_period_start"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-compute_period_start}{}}}
\subsection{Method \code{compute_period_start()}}{
Compute the period start for a specific granularity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$compute_period_start(granularity, inplace = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{granularity}}{\code{(string)} \verb{\{'hour', 'day', 'week', 'month', "year"\}}. Aggregation level.}

\item{\code{inplace}}{\code{(boolean)} If TRUE or FALSE a new instance is wanted or modifiying existing one is sufficient.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-aggregate"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-aggregate}{}}}
\subsection{Method \code{aggregate()}}{
Aggregates data according to the 'time' column at different levels. Returns datetime in the time column
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$aggregate(granularity, fun = mean, inplace = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{granularity}}{\code{(string)} \verb{\{'hour', 'day', 'month', 'year', 'week'\}} Aggregation level.}

\item{\code{fun}}{\code{(callable)} Aggregation function. 'mean' by default.}

\item{\code{inplace}}{\code{(boolean)} If TRUE or FALSE a new instance is wanted or modifiying existing one is sufficient.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-groupby"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-groupby}{}}}
\subsection{Method \code{groupby()}}{
Groupby data according to the 'time' column at different levels. Doesn't return datetime in time column
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$groupby(granularity, fun = mean)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{granularity}}{\code{(string)} \verb{\{'hour', 'wday', 'month', 'week', 'year'\}}. Aggregation level.}

\item{\code{fun}}{\code{(callable)} Aggregation function. 'mean' by default.}

\item{\code{inplace}}{\code{(boolean)} If TRUE or FALSE a new instance is wanted or modifiying existing one is sufficient.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-select"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-select}{}}}
\subsection{Method \code{select()}}{
Select variables from timeseries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$select(variables, inplace = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variables}}{\verb{(string vector)} The variables you want to select into the TimeseriesDT.}

\item{\code{inplace}}{\code{(boolean)} If TRUE or FALSE a new instance is wanted or modifiying existing one is sufficient.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-remove_duplicated"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-remove_duplicated}{}}}
\subsection{Method \code{remove_duplicated()}}{
Remove duplicated values from timeseries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$remove_duplicated(variables = "time", inplace = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variables}}{\verb{(string vector)} Default : 'time'. The variables on which you want to compare uniqueness}

\item{\code{inplace}}{\code{(boolean)} If TRUE or FALSE a new instance is wanted or modifiying existing one is sufficient.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-assign"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-assign}{}}}
\subsection{Method \code{assign()}}{
A vector to your timeseries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$assign(name, vector, inplace = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{(string)} The name of your new column}

\item{\code{vector}}{\code{(vector)} The vector to add to your timeseries}

\item{\code{inplace}}{\code{(boolean)} If TRUE or FALSE a new instance is wanted or modifiying existing one is sufficient.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-merge"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-merge}{}}}
\subsection{Method \code{merge()}}{
Merge TimeseriesDT with an other data.table or TimeseriesDT
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$merge(
  timeseries_to_merge,
  by = "time",
  how = "inner",
  suffixes = c(".x", ".y"),
  inplace = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{by}}{\verb{(string vector)} Default: 'time'. Column to merge on}

\item{\code{how}}{\code{(string)} Default : 'inner'. Either inner left or all}

\item{\code{suffixes}}{\verb{(string vector)}Suffixes to add in case of overlapping columns names}

\item{\code{inplace}}{\code{(boolean)} If TRUE or FALSE a new instance is wanted or modifiying existing one is sufficient.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-remove_variables"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-remove_variables}{}}}
\subsection{Method \code{remove_variables()}}{
Remove variables from timeseries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$remove_variables(variables, inplace = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variables}}{\verb{(string vector)} Variables to remove from timeseries}

\item{\code{inplace}}{\code{(boolean)} If TRUE or FALSE a new instance is wanted or modifiying existing one is sufficient.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-align"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-align}{}}}
\subsection{Method \code{align()}}{
Align timeseries to ensure same length. Returns input aligned
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$align(to_align, second_to_align = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{to_align}}{\code{(data.table | TimeseriesDT)} The timeseries to align with}

\item{\code{second_to_align}}{\code{(data.table | TimeseriesDT)} A third timeseries to align with}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-compute_degre_days"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-compute_degre_days}{}}}
\subsection{Method \code{compute_degre_days()}}{
Compute degree days using a threshold for heating or cooling days
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$compute_degre_days(
  temperature_column = NULL,
  all = TRUE,
  cooling = FALSE,
  threshold_cooling = 18,
  threshold_heating = 15,
  inplace = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{temperature_column}}{\code{(string)} The temperature column}

\item{\code{all}}{\code{(boolean)} Default: TRUE If TRUE, heating and cooling degree days are computed}

\item{\code{cooling}}{\code{(boolean)} If TRUE, cooling is the only to be computed}

\item{\code{threshold_cooling}}{\code{(float)} Default: 18. The threshold value for cooling}

\item{\code{threshold_heating}}{\code{(float)} Default: 15. The threshold value for heating}

\item{\code{inplace}}{\code{(boolean)} If TRUE or FALSE a new instance is wanted or modifiying existing one is sufficient.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-get_granularity"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-get_granularity}{}}}
\subsection{Method \code{get_granularity()}}{
Returns the granularity of your timeseries in the specified unit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$get_granularity(unit = "hour")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unit}}{\code{(string)} The unit of time you want the granularity. Default is hour}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-rename"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-rename}{}}}
\subsection{Method \code{rename()}}{
Renames columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$rename(old_cols, new_cols, inplace = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{old_cols}}{\verb{(string vector)} Old columns names}

\item{\code{new_cols}}{\verb{(string vector)} New columns names}

\item{\code{inplace}}{\code{(boolean)} If TRUE or FALSE a new instance is wanted or modifiying existing one is sufficient.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-filter_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-filter_dataset}{}}}
\subsection{Method \code{filter_dataset()}}{
Returns a filtered dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$filter_dataset(
  y_shifted,
  var,
  var_shifted,
  threshold,
  q_max = 0.8,
  q_min = 0.2,
  IC_width = 1.5,
  inferior = TRUE,
  inplace = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{\code{(string)} Explanatory variables to filter on}

\item{\code{var_shifted}}{\code{(string)} Explanatory variables shifted to filter on}

\item{\code{threshold}}{\code{(float)} Threshold for this explanatory variables}

\item{\code{q_max}}{\code{(float)} Quantile maximal for filtering inputs}

\item{\code{q_min}}{\code{(float)} Quantile minimum for filtering inputs}

\item{\code{IC_width}}{\code{(float)} Interval confidence to filder input}

\item{\code{inferior}}{\code{(boolean)} If TRUE or FALSE the var has to be inferior of the threshold}

\item{\code{inplace}}{\code{(boolean)} If TRUE or FALSE a new instance is wanted or modifiying existing one is sufficient.}

\item{\code{y}}{\code{(string)} Response variable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-export"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-export}{}}}
\subsection{Method \code{export()}}{
Export your timeseries to file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$export(path, as_data_table = TRUE, format = "csv")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{\code{(string)} A valid path to store the timeseries}

\item{\code{as_data_table}}{\code{(boolean)} Whether you want to export the object itself or the data.}

\item{\code{format}}{\code{(string)} \verb{\{'csv', 'rds'\}} The format to export. Optional if self = TRUE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesDT-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesDT-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesDT$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
