% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_model.R
\docType{class}
\name{TimeseriesModel}
\alias{TimeseriesModel}
\title{TimeseriesModel}
\description{
An R6 class which stands as the model structure for each climatic correction model
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{formula}}{\code{(formula)} the formula object representing the fit wanted (e.g y ~ temperature)}

\item{\code{model}}{\code{(Any)} a model object obtained such as GAM, prophet}

\item{\code{smoothers}}{\code{(MultiSmoother)} a MultiSmoother object to handle smoothing variables.}

\item{\code{by_instant}}{\code{(list)} Hyperparameter of the TimeseriesModel to take into account a fit by an instant granularity or not}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeseriesModel-new}{\code{TimeseriesModel$new()}}
\item \href{#method-TimeseriesModel-check_timeseries}{\code{TimeseriesModel$check_timeseries()}}
\item \href{#method-TimeseriesModel-get_formula}{\code{TimeseriesModel$get_formula()}}
\item \href{#method-TimeseriesModel-get_model}{\code{TimeseriesModel$get_model()}}
\item \href{#method-TimeseriesModel-get_smoothers}{\code{TimeseriesModel$get_smoothers()}}
\item \href{#method-TimeseriesModel-fit}{\code{TimeseriesModel$fit()}}
\item \href{#method-TimeseriesModel-predict}{\code{TimeseriesModel$predict()}}
\item \href{#method-TimeseriesModel-cv_predict}{\code{TimeseriesModel$cv_predict()}}
\item \href{#method-TimeseriesModel-export}{\code{TimeseriesModel$export()}}
\item \href{#method-TimeseriesModel-clone}{\code{TimeseriesModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesModel$new(formula, by_instant, granularity, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesModel-check_timeseries"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesModel-check_timeseries}{}}}
\subsection{Method \code{check_timeseries()}}{
Check if dataset in input has the variables in formula and had variables if possible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesModel$check_timeseries(X, is_fitting = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{\code{(data.table | TimeseriesDT)} The data to test}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
X \code{(data.table | TimeseriesDT)} The data tested and with calendar or shifted inputs if needed.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesModel-get_formula"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesModel-get_formula}{}}}
\subsection{Method \code{get_formula()}}{
Get the formula of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesModel$get_formula()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesModel-get_model"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesModel-get_model}{}}}
\subsection{Method \code{get_model()}}{
Get the model itself
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesModel$get_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesModel-get_smoothers"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesModel-get_smoothers}{}}}
\subsection{Method \code{get_smoothers()}}{
Get the smoothers of the MultiSmoother object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesModel$get_smoothers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesModel-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the model using the TimeseriesModel object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesModel$fit(outputs, inputs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outputs}}{\code{(data.table | TimeseriesDT)} The output/response data}

\item{\code{inputs}}{\code{(data.table | TimeseriesDT)} The input data to fit on.}

\item{\code{fold_varname}}{\code{(string)} the name of the variable in \code{inputs} to define CV folds from}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesModel-predict"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesModel-predict}{}}}
\subsection{Method \code{predict()}}{
Predict the model and make the climatic correction using the an Operator Object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesModel$predict(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{\code{(data.table | TimeseriesDT)} The timeseries data to make prediction on}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(vector)} The output timeseries as a vector from the model prediction
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesModel-cv_predict"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesModel-cv_predict}{}}}
\subsection{Method \code{cv_predict()}}{
Make CV predictions from the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesModel$cv_predict(outputs, inputs, fold_varname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outputs}}{\code{(data.table | TimeseriesDT)} The output/response data}

\item{\code{inputs}}{\code{(data.table | TimeseriesDT)} The input data}

\item{\code{fold_varname}}{\code{(string)} the name of the variable in \code{inputs} to define CV folds from}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(vector)} The output timeseries as a vector from the model CV prediction
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesModel-export"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesModel-export}{}}}
\subsection{Method \code{export()}}{
Export the TimeseriesModel in RDS.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesModel$export(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{\code{(string)} Path file to write the rds file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
