% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_std_model.R
\docType{class}
\name{TimeseriesStdModel}
\alias{TimeseriesStdModel}
\title{TimeseriesStdModel}
\description{
An R6 class which stands as the model structure for each conditional standard deviation timeseries model
}
\section{Super class}{
\code{\link[CorrClim:TimeseriesModel]{CorrClim::TimeseriesModel}} -> \code{TimeseriesStdModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{formula}}{\code{(formula)} the formula object representing the fit wanted (e.g value ~ temperature)}

\item{\code{model}}{\code{(Any)} a model object obtained such as GAM, prophet}

\item{\code{smoothers}}{\code{(Smoother)} a Smoother object}

\item{\code{conditional_expectation_model}}{\code{(TimeseriesModel)} the model to compute conditional expectation from}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{formula}}{\code{(formula)} the formula object representing the fit wanted (e.g value ~ temperature)}

\item{\code{model}}{\code{(Any)} a model object obtained such as GAM, prophet}

\item{\code{smoothers}}{\code{(Smoother)} a Smoother object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeseriesStdModel-new}{\code{TimeseriesStdModel$new()}}
\item \href{#method-TimeseriesStdModel-fit}{\code{TimeseriesStdModel$fit()}}
\item \href{#method-TimeseriesStdModel-predict}{\code{TimeseriesStdModel$predict()}}
\item \href{#method-TimeseriesStdModel-clone}{\code{TimeseriesStdModel$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="check_timeseries"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-check_timeseries'><code>CorrClim::TimeseriesModel$check_timeseries()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="cv_predict"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-cv_predict'><code>CorrClim::TimeseriesModel$cv_predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="export"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-export'><code>CorrClim::TimeseriesModel$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_formula"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_formula'><code>CorrClim::TimeseriesModel$get_formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_model"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_model'><code>CorrClim::TimeseriesModel$get_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CorrClim" data-topic="TimeseriesModel" data-id="get_smoothers"><a href='../../CorrClim/html/TimeseriesModel.html#method-TimeseriesModel-get_smoothers'><code>CorrClim::TimeseriesModel$get_smoothers()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesStdModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesStdModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesStdModel$new(formula, by_instant, granularity, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesStdModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesStdModel-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the conditional variance model based on the conditional expectation one
Relies on CV squared residuals computed from the latter to do so
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesStdModel$fit(outputs, inputs, fold_varname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outputs}}{\code{(data.table | TimeseriesDT)} The output/response data}

\item{\code{inputs}}{\code{(data.table | TimeseriesDT)} The input data with data for both conditional
expectation and conditional variance models}

\item{\code{fold_varname}}{\code{(string)} the name of the variable in \code{inputs} to define CV folds from}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesStdModel-predict"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesStdModel-predict}{}}}
\subsection{Method \code{predict()}}{
Predict the conditional standard deviation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesStdModel$predict(inputs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputs}}{\code{(data.table | TimeseriesDT)} The timeseries data to make prediction on}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{(vector)} The output timeseries as a vector from the model prediction
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeseriesStdModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeseriesStdModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeseriesStdModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
